zmiana pola id z int na long

+++ schoolboard-image-manager/src/dataaccess/LabelDAO.java
@@ -3,7 +3,7 @@ package dataaccess;
 import imagemanager.model.Label;
 import imagemanager.model.SourceImage;
 
-public interface LabelDAO extends GenericDAO<Label, Integer> {
+public interface LabelDAO extends GenericDAO<Label, Long> {
 	
 	public void assignLabel(Label label, SourceImage image);
 	public void unassignLabel(Label label, SourceImage image);
